Move News.vue, Statistics.vue, Syntax.vue to penalty/sidebar subfolder and update imports
