+++ web/components/penalty/sidebar/Statistics.vue
