+++ web/components/penalty/sidebar/Syntax.vue
